Moved getStaticReferences to contractHelper
